Add Title to Pie and Doughnut Chart

+++ smartgraph/src/main/java/originator/ailin/com/smartgraph/title/Title.java
@@ -9,16 +9,31 @@ import android.graphics.Paint;
 public class Title {
     private Canvas mCanvas;
     private Paint mPaint;
-    private int mLeft, mBottom;
+    private float mLeft, mBottom;
 
-    public Title(Canvas mCanvas, Paint mPaint, int mLeft, int mBottom) {
+    /**
+     *
+     * @param mCanvas
+     * @param mPaint
+     * @param mLeft
+     * @param mBottom
+     */
+    public Title(Canvas mCanvas, Paint mPaint, float mLeft, float mBottom) {
         this.mCanvas = mCanvas;
         this.mPaint = mPaint;
         this.mLeft = mLeft;
         this.mBottom = mBottom;
     }
 
-    public void drawTitle(String titleText, int titleSize, int titleColor, int maxWidth, int maxHeight) {
+    /**
+     *
+     * @param titleText
+     * @param titleSize
+     * @param titleColor
+     * @param maxWidth
+     * @param maxHeight
+     */
+    public void drawTitle(String titleText, int titleSize, int titleColor, float maxWidth, float maxHeight) {
         mPaint.setStyle(Paint.Style.FILL);
         mPaint.setStrokeWidth(1);
         mPaint.setTextSize(titleSize);
